Disabled some stuff
Recursive probing is way too much for /lib and the $(which ffmpeg) search should in theory work most of the time if there is an installation. Since this is a build script where a knowledgeable user presumably has the authority over where it is run, this should not be a problem.