Reversed handling method
So far, MediaReceiver has established its own track and requested it. This is not very useful because we can only indicate that we would like a video and UE ultimately has the choice of media. The logic here is kind of leftover from when PX1 used to accept control offers for data con without video.