Patched build errors for unix
Windows builds have colocated libraries which make the linking easier by default, but we might need to adapt here as this might cause future issues, but for now just the unix build was adapted. We can use vcpkg on unix as well -- it just makes some parts easier when built in weird contexts (such as certain libraries missing in the module system).